refactor(backend): extract file crud module
Move the seven file functions out of crud/__init__.py into crud/file.py,
following the split already done for comment, token, metric and others.
Consumers import the module directly; nothing is re-exported.

All seven have callers, so nothing was dropped. Removing the block left
func unused at module scope in crud/__init__.py, which turned the local
"from sqlalchemy import func" inside get_user_by_email into an F811, so
the module-level import of func goes too.

telemetry.py and routers/file.py no longer need "from rhesis.backend.app
import crud" at all; tracing.py and conversation_linking.py swap their
function-local "crud as _crud" imports for the file module.
