+++ apps/backend/src/rhesis/backend/app/routers/test_result.py
@@ -14,6 +14,7 @@ from rhesis.backend.app.auth.capabilities import Permission
 from rhesis.backend.app.auth.principal import resolve_principal_from_request
 from rhesis.backend.app.auth.rbac import authorize_object, project_id_from_scope
 from rhesis.backend.app.auth.user_utils import require_current_user_or_token
+from rhesis.backend.app.crud import file as file_crud
 from rhesis.backend.app.dependencies import (
     get_tenant_context,
     get_tenant_db_session,
@@ -593,4 +594,6 @@ def list_test_result_files(
 ):
     """List output files attached to a test result."""
     organization_id, user_id = tenant_context
-    return crud.get_files_for_entity(db, test_result_id, "TestResult", organization_id, user_id)
+    return file_crud.get_files_for_entity(
+        db, test_result_id, "TestResult", organization_id, user_id
+    )
